--- src/routes/graphql/api/IUserType.ts
@@ -1,9 +1,0 @@
-export interface IUserType {
-  id: string;
-  name: string;
-  balance: number;
-  profile: unknown;
-  posts: unknown[];
-  userSubscribedTo: IUserType[];
-  subscribedToUser: IUserType[];
-}
